refactor: remove unused autoprobe_analysis module files and update import paths for consistency
